move colorToRgba into wrappers
