+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/styles/area.test.ts
@@ -7,7 +7,6 @@
  */
 
 import * as commonColors from '../../../../../common/color_calcs';
-import { colorToRgba } from '../../../../../common/color_calcs';
 import { Fill } from '../../../../../geoms/types';
 import { getMockCanvas, getMockCanvasContext2D, MockStyles } from '../../../../../mocks';
 import * as common from '../../../../../utils/common';
@@ -15,6 +14,7 @@ import { getTextureStyles } from '../../../utils/texture';
 import { buildAreaStyles } from './area';
 
 import 'jest-canvas-mock';
+import { colorToRgba } from '../../../../../common/color_library_wrappers';
 
 jest.mock('../../../utils/texture');
 jest.spyOn(common, 'getColorFromVariant');
