+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/styles/area.ts
@@ -6,8 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { colorToRgba } from '../../../../../common/color_calcs';
-import { OpacityFn, overrideOpacity } from '../../../../../common/color_library_wrappers';
+import { colorToRgba, OpacityFn, overrideOpacity } from '../../../../../common/color_library_wrappers';
 import { Fill } from '../../../../../geoms/types';
 import { Color, ColorVariant, getColorFromVariant } from '../../../../../utils/common';
 import { GeometryStateStyle, AreaStyle } from '../../../../../utils/themes/theme';
